requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
colorama==0.4.4
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.3
django-filter==2.4.0
django-heroku==0.3.1
django-rest==0.8.6
djangorestframework==3.12.2
filelock==3.0.12
gunicorn==20.0.4
isort==5.6.4
lazy-object-proxy==1.4.3
Markdown==3.3.3
mccabe==0.6.1
Pillow==8.0.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
virtualenv==20.2.1
whitenoise==5.2.0
wrapt==1.12.1
